+++ src/Interfaces/GitMeHookedConfig.ts
@@ -6,6 +6,7 @@ interface Command {
   exec: string;
   name?: string;
   silence?: boolean;
+  config?: object;
 }
 
 interface GitMeHookedConfig {